+++ src/structure/home/home.component.ts
@@ -1,5 +1,9 @@
 import { Component, OnInit } from '@angular/core';
-import { ActivatedRoute} from '@angular/router';
+import { getDatabase, ref, onValue } from 'firebase/database';
+import { ActivatedRoute, Router} from '@angular/router';
+import { initializeApp } from 'firebase/app';
+import { environment } from 'src/environments/environment';
+
 
 @Component({
   selector: 'app-home',
@@ -8,18 +12,17 @@ import { ActivatedRoute} from '@angular/router';
 })
 export class HomeComponent implements OnInit {
 
-  constructor(private route:ActivatedRoute) { }
+  constructor(private route:ActivatedRoute, private router:Router) { }
 
   username:any= ""
-  public w:any;
+  email:any = ""
+  // public w:any;
 
 
   ngOnInit(): void {
-    this.route.queryParamMap.subscribe(params => this.username = params.get("user"));
-
-    this.w = window.innerWidth;
-
-
+    this.route.queryParamMap.subscribe(params => this.email = params.get("email"));
+    this.fetchUsername(this.email)
+    // this.w = window.innerWidth;
   }
 
 
@@ -28,4 +31,25 @@ export class HomeComponent implements OnInit {
   {
     console.log("Changed")
   }
+
+  go_to_profile() {
+    this.router.navigate(['/profile'], 
+    {
+      queryParams : { email: this.email}
+    })
+  }
+
+  app = initializeApp(environment.firebase)
+
+  fetchUsername(em : string) {
+  const realdb = getDatabase(this.app)
+  em = em.replace(".", "")
+  const realRef = ref(realdb, '/users/' + em);
+  onValue(realRef, (snapshot) => {
+    const data = snapshot.val();
+    this.username = data.username    
+  });
+}
+
+
 }