+++ src/users/register/register.component.ts
@@ -17,29 +17,22 @@ export class RegisterComponent implements OnInit {
   constructor(private router: Router, private route:ActivatedRoute) { }
 
   ngOnInit(): void {
+    console.log("omkarelectric892@gmail.com".replace('.',''))
   }
-
 
   app = initializeApp(environment.firebase)
 
   auth = getAuth(this.app)
-
   
-
-
-  register(fullname:any, uname:any, email:any, password:any) {
-    const name = fullname
-    const loginuname = uname
-    const loginEmail = email
-    const loginPass = password
-
+  register(fullname:any, uname:any, email:any, password:any) {    
+    let keymail = email.replace('.','')
     const database = getDatabase()
-    let userRef = ref(database, 'users/')
-    let newUser = push(userRef)
-    set(newUser, {
-      name: name,
-      username: loginuname,
-      email: loginEmail      
+    let userRef = ref(database, 'users/' + keymail)
+    //let newUser = push(userRef)
+    set(userRef, {
+      name: fullname,
+      username: uname,
+      email: email      
     }).then(() => {
       // Data saved successfully!
       console.log("Data updated in realtime database")
@@ -50,7 +43,7 @@ export class RegisterComponent implements OnInit {
     });
 
 
-    createUserWithEmailAndPassword(this.auth, loginEmail, loginPass)
+    createUserWithEmailAndPassword(this.auth, email, password)
   .then((userCredential) => {
 
     const user = userCredential.user;